Add files via upload
Input code contains functions to read and verify user input code and return the code for future use. For the input to work it is necessary to download the internal-data excel file and the input txt file and replace filepaths leading to the files in your repository.

Functions contain error conditions to check whether the code is valid and whether the code refers to a date that has not yet occurred.